Improve structure and documentation of CurrentFilters connector
